outsourced collection and computed shared tests
