+++ packages/core/tests/unit/computed/index.test.ts
@@ -1,0 +1,86 @@
+import {
+  Agile,
+  assignSharedAgileInstance,
+  Computed,
+  createComputed,
+} from '../../../src';
+import { LogMock } from '../../helper/logMock';
+
+jest.mock('../../../src/computed/computed');
+
+describe('Computed Index', () => {
+  let sharedAgileInstance: Agile;
+
+  beforeEach(() => {
+    LogMock.mockLogs();
+
+    sharedAgileInstance = new Agile();
+    assignSharedAgileInstance(sharedAgileInstance);
+
+    jest.clearAllMocks();
+  });
+
+  describe('createComputed function tests', () => {
+    const ComputedMock = Computed as jest.MockedClass<typeof Computed>;
+    const computedFunction = () => {
+      // empty
+    };
+
+    beforeEach(() => {
+      ComputedMock.mockClear();
+    });
+
+    it('should create Computed with the shared Agile Instance (default config)', () => {
+      const response = createComputed(computedFunction, ['dummyDep' as any]);
+
+      expect(response).toBeInstanceOf(Computed);
+      expect(ComputedMock).toHaveBeenCalledWith(
+        sharedAgileInstance,
+        computedFunction,
+        {
+          computedDeps: ['dummyDep' as any],
+        }
+      );
+    });
+
+    it('should create Computed with the shared Agile Instance (specific config)', () => {
+      const computedConfig = {
+        key: 'jeff',
+        isPlaceholder: false,
+        computedDeps: ['dummyDep' as any],
+        autodetect: true,
+      };
+
+      const response = createComputed(computedFunction, computedConfig);
+
+      expect(response).toBeInstanceOf(Computed);
+      expect(ComputedMock).toHaveBeenCalledWith(
+        sharedAgileInstance,
+        computedFunction,
+        computedConfig
+      );
+    });
+
+    it('should create Computed with a specified Agile Instance (specific config)', () => {
+      const agile = new Agile();
+      const computedConfig = {
+        key: 'jeff',
+        isPlaceholder: false,
+        computedDeps: ['dummyDep' as any],
+        autodetect: true,
+      };
+
+      const response = createComputed(computedFunction, {
+        ...computedConfig,
+        ...{ agileInstance: agile },
+      });
+
+      expect(response).toBeInstanceOf(Computed);
+      expect(ComputedMock).toHaveBeenCalledWith(
+        agile,
+        computedFunction,
+        computedConfig
+      );
+    });
+  });
+});
