+++ packages/core/tests/unit/shared.test.ts
@@ -9,9 +9,6 @@ import {
 } from '../../src';
 import { LogMock } from '../helper/logMock';
 
-jest.mock('../../src/collection/collection');
-jest.mock('../../src/computed/computed');
-
 describe('Shared Tests', () => {
   let sharedAgileInstance: Agile;
 
@@ -31,110 +28,6 @@ describe('Shared Tests', () => {
       assignSharedAgileInstance(newAgileInstance);
 
       expect(shared).toBe(newAgileInstance);
-    });
-  });
-
-  describe('createCollection function tests', () => {
-    const CollectionMock = Collection as jest.MockedClass<typeof Collection>;
-
-    beforeEach(() => {
-      CollectionMock.mockClear();
-    });
-
-    it('should create Collection with the shared Agile Instance', () => {
-      const collectionConfig = {
-        selectors: ['test', 'test1'],
-        groups: ['test2', 'test10'],
-        defaultGroupKey: 'frank',
-        key: 'myCoolCollection',
-      };
-
-      const collection = createCollection(collectionConfig);
-
-      expect(collection).toBeInstanceOf(Collection);
-      expect(CollectionMock).toHaveBeenCalledWith(
-        sharedAgileInstance,
-        collectionConfig
-      );
-    });
-
-    it('should create Collection with a specified Agile Instance', () => {
-      const agile = new Agile();
-      const collectionConfig = {
-        selectors: ['test', 'test1'],
-        groups: ['test2', 'test10'],
-        defaultGroupKey: 'frank',
-        key: 'myCoolCollection',
-      };
-
-      const collection = createCollection(collectionConfig, agile);
-
-      expect(collection).toBeInstanceOf(Collection);
-      expect(CollectionMock).toHaveBeenCalledWith(agile, collectionConfig);
-    });
-  });
-
-  describe('createComputed function tests', () => {
-    const ComputedMock = Computed as jest.MockedClass<typeof Computed>;
-    const computedFunction = () => {
-      // empty
-    };
-
-    beforeEach(() => {
-      ComputedMock.mockClear();
-    });
-
-    it('should create Computed with the shared Agile Instance (default config)', () => {
-      const response = createComputed(computedFunction, ['dummyDep' as any]);
-
-      expect(response).toBeInstanceOf(Computed);
-      expect(ComputedMock).toHaveBeenCalledWith(
-        sharedAgileInstance,
-        computedFunction,
-        {
-          computedDeps: ['dummyDep' as any],
-        }
-      );
-    });
-
-    it('should create Computed with the shared Agile Instance (specific config)', () => {
-      const computedConfig = {
-        key: 'jeff',
-        isPlaceholder: false,
-        computedDeps: ['dummyDep' as any],
-        autodetect: true,
-      };
-
-      const response = createComputed(computedFunction, computedConfig);
-
-      expect(response).toBeInstanceOf(Computed);
-      expect(ComputedMock).toHaveBeenCalledWith(
-        sharedAgileInstance,
-        computedFunction,
-        computedConfig
-      );
-    });
-
-    it('should create Computed with a specified Agile Instance (specific config)', () => {
-      const agile = new Agile();
-      const computedConfig = {
-        key: 'jeff',
-        isPlaceholder: false,
-        computedDeps: ['dummyDep' as any],
-        autodetect: true,
-      };
-
-      const response = createComputed(computedFunction, {
-        ...computedConfig,
-        ...{ agileInstance: agile },
-      });
-
-      expect(response).toBeInstanceOf(Computed);
-      expect(ComputedMock).toHaveBeenCalledWith(
-        agile,
-        computedFunction,
-        computedConfig
-      );
     });
   });
 });
